refactor WorkflowResult to results/ dir to support more more results releated objects in future
