+++ src/ploigos_step_runner/results/__init__.py
@@ -3,3 +3,4 @@
 
 from ploigos_step_runner.results.step_result import StepResult
 from ploigos_step_runner.results.step_result_artifact import StepResultArtifact
+from ploigos_step_runner.results.workflow_result import WorkflowResult
