+++ src/ploigos_step_runner/results/workflow_result.py
@@ -5,9 +5,8 @@ import os
 import pickle
 
 import yaml
-
 from ploigos_step_runner.exceptions import StepRunnerException
-from ploigos_step_runner import StepResult
+from ploigos_step_runner.results import StepResult
 from ploigos_step_runner.utils.dict import deep_merge
 from ploigos_step_runner.utils.file import create_parent_dir
