+++ src/ploigos_step_runner/step_runner.py
@@ -4,9 +4,9 @@ import os
 
 from ploigos_step_runner.config.config import Config
 from ploigos_step_runner.exceptions import StepRunnerException
+from ploigos_step_runner.results import WorkflowResult
 from ploigos_step_runner.step_implementer import StepImplementer
 from ploigos_step_runner.utils.reflection import import_and_get_class
-from ploigos_step_runner.workflow_result import WorkflowResult
 
 
 class StepRunner:
